Design review items (part 1), added basic derived LoadTestConfigPayload model
