+++ src/LodeRunner.API/src/Controllers/LoadTestConfigController.cs
@@ -1,14 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using System;
-using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using AutoMapper;
-using LodeRunner.API.Interfaces;
 using LodeRunner.API.Middleware;
-using LodeRunner.API.Models;
 using LodeRunner.Core.Models;
 using LodeRunner.Data.Interfaces;
 using Microsoft.AspNetCore.Mvc;
@@ -45,18 +41,19 @@ namespace LodeRunner.API.Controllers
             Summary = "Creates a new LoadTestConfig item",
             Description = "Requires Load Test Config payload",
             OperationId = "CreateLoadTestConfig")]
-        public IActionResult CreateLoadTestConfig([FromBody, SwaggerRequestBody("The load test config payload", Required = true)] LoadTestConfig loadTestConfigPayload, [FromServices] ILoadTestConfigService loadTestConfigService, [FromServices] CancellationTokenSource cancellationTokenSource)
+        public IActionResult CreateLoadTestConfig([FromBody, SwaggerRequestBody("The load test config payload", Required = true)] LoadTestConfigPayload loadTestConfigPayload, [FromServices] ILoadTestConfigService loadTestConfigService, [FromServices] CancellationTokenSource cancellationTokenSource)
         {
             if (cancellationTokenSource != null && cancellationTokenSource.IsCancellationRequested)
             {
                 return ResultHandler.CreateCancellationInProgressResult();
             }
 
-            LoadTestConfig loadTestConfig = autoMapper.Map<LoadTestConfig>(loadTestConfigPayload);
+            // NOTE: the Mapping configuration will create a new loadTestConfig but will ignore the Id since the property has a getter and setter.
+            LoadTestConfig newloadTestConfig = autoMapper.Map<LoadTestConfig>(loadTestConfigPayload);
 
-            if (loadTestConfig.Validate(out string errorMessage))
+            if (newloadTestConfig.Validate(out string errorMessage))
             {
-                var insertedLoadTestConfig = loadTestConfigService.Post(loadTestConfig, cancellationTokenSource.Token).Result;
+                var insertedLoadTestConfig = loadTestConfigService.Post(newloadTestConfig, cancellationTokenSource.Token).Result;
 
                 if (insertedLoadTestConfig != null)
                 {
